deleted jwplayer, created java utils helper file
