+++ ios/ios/PlatformContext_iOS.swift
@@ -17,16 +17,6 @@ protocol PlatformContext_iOSDelegate : class {
 
 class PlatformContext_iOS : NSObject {
     
-    
-    static func javaList(toList list: JavaUtilList) -> [Any] {
-        var array = [Any]()
-        for i in 0..<list.size() {
-            array.append(list.getWith(i))
-        }
-        return array
-    }
-    
-    
     var isDayTwo: Bool = false
     var dateFormatter: DateFormatter!
     var timeFormatter: DateFormatter!
@@ -34,8 +24,6 @@ class PlatformContext_iOS : NSObject {
     var conferenceDays: [DPRESDaySchedule]?
     
     weak var reloadDelegate: PlatformContext_iOSDelegate?
-    
-    
     
     lazy var storageDir: String = {
         let paths = NSSearchPathForDirectoriesInDomains(FileManager.SearchPathDirectory.documentDirectory, FileManager.SearchPathDomainMask.userDomainMask, true)
@@ -48,7 +36,7 @@ class PlatformContext_iOS : NSObject {
         
         let index = self.isDayTwo ? 1 : 0
         if let days = conferenceDays, days.count > index, let objectArray = days[index].getHourHolders() {
-            array.append(contentsOf: PlatformContext_iOS.convertiOSObjectArrayToArray(objArray: objectArray) as! [DPRESHourBlock])
+            array.append(contentsOf: JavaUtils.convertiOSObjectArrayToArray(objArray: objectArray) as! [DPRESHourBlock])
         }
         if dateFormatter == nil {
             dateFormatter = DateFormatter()
@@ -58,17 +46,8 @@ class PlatformContext_iOS : NSObject {
     }
     
     func getSpeakersArray(from networkEvent: DDATEvent) -> [Any] {
-        return PlatformContext_iOS.javaList(toList: networkEvent.getSpeakerList())
+        return JavaUtils.javaList(toList: networkEvent.getSpeakerList())
     }
-    
-//    fileprivate func formatSpeakersString(from array: [DDATEventSpeaker]) -> String {
-//        var speakerNames = [String]()
-//        //TODO: Fix this
-////        for speaker in array {
-////            speakerNames.append(speaker.getUserAccount().getName())
-////        }
-//        return speakerNames.joined(separator: ", ")
-//    }
     
     func getEventTime(startTime: NSString, andEnd endTime: NSString) -> String {
         var startLoc = 7
@@ -101,16 +80,6 @@ class PlatformContext_iOS : NSObject {
         hourBlocks.removeAll()
         hourBlocks.append(contentsOf: hourBlocksArray)
     }
-    
-    //TODO move somewhere else
-    static func convertiOSObjectArrayToArray(objArray: IOSObjectArray) -> [Any] {
-        var array = [Any]()
-        for i in 0..<objArray.length() {
-            array.append(objArray.object(at: UInt(i)))
-        }
-        return array
-    }
-    
 }
 
 extension PlatformContext_iOS : UITableViewDataSource {
@@ -172,10 +141,9 @@ extension PlatformContext_iOS : DPRESScheduleDataViewModel_Host {
     
         func loadCallback(withDPRESDayScheduleArray daySchedules: IOSObjectArray!) {
             hourBlocks = [DPRESHourBlock]()
-            conferenceDays = PlatformContext_iOS.convertiOSObjectArrayToArray(objArray: daySchedules) as! [DPRESDaySchedule]
+            conferenceDays = JavaUtils.convertiOSObjectArrayToArray(objArray: daySchedules) as! [DPRESDaySchedule]
             updateTableData()
             reloadDelegate?.reloadTableView()
         }
-    
 }
